wip: change token format to "ast-style" dict
That is, I use a dict where a _kind_ (that is, a type) is specified,
and any needed fields (such as the spelling of an identifier) are
tagged on.

This is WIP, because all of the tests should fail, except for the
tokenization ones (that is, 'test_tokenize_lambda.py').