chore: move project out of 'python' directory
I had been first thinking of doing this in Go, but then decided to try
my hand at Python and so had made a separate subdirectory for
it. However, it looks like I'm completely moving forward with Python,
and so there's no longer a need to house any Python code in a
subdirectory; it call all live at the top level, like a normal
project.