fix: pass no_genv tests using Exception-returning style
1. Special helper function for getting a token at a given index, to
make tracking "falling of the edge of the world" errors easier.

2. Improved comments for weird parts of the code, like our fold-left
algorithm for desugaring multi-term application syntax.

3. Refactor the unit tests to use our Exception-returning style.

4. Get rid of our old ParseError classes in favor of a simpler design
based on a single ParseError dataclass that inherits from Exception,
and uses an enum tag that basically categorizes the various kinds of
errors in a discrete, decisive manner.

Up to now, I still only intend on passing the no_genv tests.