refactor: use inheritance hierarchy to define AST
This hopefully results in cleaner, more refactorable code.

With a little rebasing, this is now an official refactor, as opposed
to a wip. But I'll leave the following (now tentatively obsolete)
notes from the past as a possibly useful historical artifact:

<ahem>

Some tests, however, fail (possibly because of a bug in how free
symbols are handled now.)

1. A good first step would be to give a str method to the AST kinds, to
make the printouts more readable.

2. Hopefully there's a way to make the label parameter optional when
defining the various ASTs, while still letting that field hang off the
parent AST class as an optional field. Maybe a post-init trick will do
here, like setting it to None, then when a field is needed, call a
method that sets it, that then returns the object (so we wouldn't need
a separate, "destructive" step to configure the object.)