fix: make i point to correct token when reporting ASSIGN_TO_LOCAL
Note that this has also been renamed from SHADOWED_FREE_NAME, since
the variable being assigned to may or may not have been declared as a
free name; the important thing to flag is whether the variable being
assigned to is already bound locally.

The idea of this error in general is that one should always use
lambda-application to assign a value to a local variable.