chore: define addition in Lambda Term code
This is the definition that first cued me that something was wrong
with my interpreter. Now it works! Only, there's something curious
about it: the order of the arguments results in terms that are unequal
in form, but have the same meaning.

Maybe implementing 'equal', another recursive function, could resolve
this point.