feat: pretty print unused lambda binders with '_'
This could be helpful for occasional debugging, since it's sometimes
useful to instrument lambda terms with a preceding '\d.', thus
delaying their evaluation by wrapping them in an abstraction. The 'd'
stands for "debug." However, since this term doesn't appear in the
wrapped expression itself, it becomes difficult to differentiate this
binder (which gets "mangled" later on using a random word) from the
other binders that appear in the expression.