feat! reinclude 'aux' parameter in term.Name
I was wrong before: what I was using as the 'depth' field inside
term.Name doesn't always have to reflect that name's depth after a
beta reduction (though it starts out with that value.) Rather, it
seems to help ID even local names.

Observe that the name-index is managed by _shift to preserve the
meaning of global names. Global name indices are defined by adding the
length of the environment - the depth - to the gamma value:

global_index = gamma + depth

When beta reducing a name, this second parameter (initialized as the
depth) achieves a value of zero, and so the index suffices as the
gamma value itself.

However, we can therefore preserve a second index, 'aux' (initialized
as the depth), and ensure that

index - aux

stays constant across beta reduction. The value 'aux' won't always
refer to the depth, but it will help identify _where_ every variable
is after beta reduction, even local ones.

Notably, 'bind' had to be updated to accomodate these new semantics.

The only thing left to do now is update the tests to use the new Name
constructor syntax, but things look good so far.