feat: define equality between Names
Let's assume we enter a definition of 'second' in the REPL:

\x.\y.x

The AST that results is

F(F(N(0, 1)))

This is because 'aux' for newly-defined names gets assigned as that
name's scope depth.

However, consider the following expression:

(apply (identity first) identity)

This gives us the following AST:

F(F(N(0, 1)))

This is OK as a definition of 'second', because the inner name still
has an index of 0, meaning it's being bound by the inner lambda, as
opposed to the outer one (or else by neither, which would make it
global.) However, my understanding is that the given name had started
out with an 'index - aux' value of -1, which has been preserved across
beta reduction.