wip: avoid infinite recursion when unwrapping thunked refs
It looks like there's a problem with how global variables are being
referenced in this scheme. It's a problem I've run into before in
other contexts.

When I evaluate

(add two three)

I get something which I suspect is structurally identical to the
expected answer of (Church numeral) five; but the global labels look
all wrong.