fix: roll back letrec-based code
As of now, I can unskip the previously lengthy tests.

I also had to include a missing definition for "not", which
unfortunately cost me more than an hour of debugging. :( This was the
reason why div expressions worked perfectly in the REPL, but
mysteriously hung in the tests.

Along those lines, I've left in some commented code that could be
useful for any future debugging later.